cleaning up the code and moving things out of the header file
